tools/requirements.txt
asteval==0.9.31
contourpy==1.1.0
cycler==0.11.0
declxml==1.1.3
exceptiongroup==1.1.3
fonttools==4.42.1
future==0.18.3
greenlet==2.0.2
iniconfig==2.0.0
isodate==0.6.1
kiwisolver==1.4.5
lmfit==1.2.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.7.3
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pluggy==1.3.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyparsing==3.1.1
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
pytest==7.4.2
pytest-qt==4.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyUserInput==0.1.11
PyYAML==6.0.1
scipy==1.11.2
seaborn==0.12.2
shiboken6==6.5.2
six==1.16.0
SQLAlchemy==1.4
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
uncertainties==3.1.7
